gafftype.[h/cpp]: Make sure function signatures match the ones from OBFFType (unsigned int vs. size_t)

+++ src/forcefields/gaff/gafftype.cpp
@@ -575,23 +575,23 @@ namespace OpenBabel {
       return valid;
     }
 
-    bool GAFFType::IsConnected(const unsigned int &idxA, const unsigned int &idxB) const
+    bool GAFFType::IsConnected(const size_t &idxA, const size_t &idxB) const
     {
       return (m_Connected.find((idxA-1)+m_numAtoms*(idxB-1))!=m_Connected.end());
     }
 
-    bool GAFFType::IsOneThree(const unsigned int &idxA, const unsigned int &idxB) const
+    bool GAFFType::IsOneThree(const size_t &idxA, const size_t &idxB) const
     {
       return (m_OneThree.find((idxA-1)+m_numAtoms*(idxB-1))!=m_OneThree.end());
     }
 
-    bool GAFFType::IsOneFour(const unsigned int &idxA, const unsigned int &idxB) const
+    bool GAFFType::IsOneFour(const size_t &idxA, const size_t &idxB) const
     {
       return (m_OneFour.find((idxA-1)+m_numAtoms*(idxB-1))!=m_OneFour.end());
     }
 
 
-    const string & GAFFType::GetAtomType(const unsigned int & idx) const
+    const string & GAFFType::GetAtomType(const size_t & idx) const
     {
       return m_atoms[idx-1];
     }
